Fix type conversion warnings in PBRefController
Changed int to NSInteger for row, column, and subjectColumn variables to match
the return types of NSTableView methods and eliminate -Wshorten-64-to-32 warnings.

+++ Classes/Controllers/PBRefController.m
@@ -262,9 +262,9 @@
 {
     
 	NSPoint location = [(PBCommitList *)tv mouseDownPoint];
-	int row = [tv rowAtPoint:location];
-	int column = [tv columnAtPoint:location];
-	int subjectColumn = [tv columnWithIdentifier:@"SubjectColumn"];
+	NSInteger row = [tv rowAtPoint:location];
+	NSInteger column = [tv columnAtPoint:location];
+	NSInteger subjectColumn = [tv columnWithIdentifier:@"SubjectColumn"];
 	if (column != subjectColumn)
 		return NO;
